- fixed all comments

+++ src/main/java/com/example/six_entities/exception/FileNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.six_entities.exception;
 
 public class FileNotFoundException extends RuntimeException {
-    public FileNotFoundException(Long id) {
+    public FileNotFoundException(Integer id) {
         super("File not found, id=" + id);
     }
 }
