+++ src/main/java/com/example/six_entities/repository/FileRepository.java
@@ -3,5 +3,5 @@ package com.example.six_entities.repository;
 import com.example.six_entities.model.File;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepository extends JpaRepository<File, Long> {
+public interface FileRepository extends JpaRepository<File, Integer> {
 }
